+++ assets/javascripts/kitten/components/icons/studing-icon.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export const StudingIcon = props => {
+  <svg xmlns="http://www.w3.org/2000/svg"
+       viewBox="0 0 11 11"
+       { ...props }>
+    <circle cx="5.5" cy="5.5" r="5" fill="#fff" stroke="#333"/>
+    <path d="M6 5h1.5v1H5V2.5h1z" fill="#333"/>
+  </svg>
+
+}